fix new inventory API
